+++ Prueba.Web/Services/Interfaces/IOperationsServices.cs
@@ -6,6 +6,6 @@ namespace Prueba.Web.Repositories.Interfaces
     public interface IOperationsServices
     {
         IEnumerable<Operation> GetAll();
-        Task<ServiceResult<OperationResponseDto>> AddAsync(int valueA, int valueB);
+        Task<ServiceResult<OperationResponseDto>> AddAsync(long valueA, long valueB);
     }
 }
